Add Protocol - InputPasswordComponentProtocol
