--- Projects/App/Sources/Present/Auth/SignUp/RegistrationPassword/Component/inputPassword.swift
@@ -1,9 +1,0 @@
-//
-//  inputPassword.swift
-//  Choice
-//
-//  Created by 민도현 on 2023/06/08.
-//  Copyright © 2023 dohyeon. All rights reserved.
-//
-
-import Foundation
